refactor: separar tabela de remessas em arquivos organizados
- Criar remessa-columns.tsx com definições das colunas
- Criar remessa-table.tsx com componente da tabela
- Separar responsabilidades entre colunas e lógica da tabela
- Melhorar organização e reutilização de componentes
- Atualizar exportações no index.ts
