+++ src/features/remessa/components/tables/index.ts
@@ -1,0 +1,2 @@
+export { RemessaTable } from './remessa-table';
+export { remessaColumns } from './remessa-columns';
