+++ src/features/remessa/components/tables/remessa-columns.tsx
@@ -1,29 +1,5 @@
-'use client';
-
-import * as React from 'react';
-import {
-	ColumnDef,
-	ColumnFiltersState,
-	flexRender,
-	getCoreRowModel,
-	getFilteredRowModel,
-	getPaginationRowModel,
-	getSortedRowModel,
-	SortingState,
-	useReactTable,
-} from '@tanstack/react-table';
-
+import { ColumnDef } from '@tanstack/react-table';
 import { Button } from '@/components/ui/button';
-import {
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
-} from '@/components/ui/table';
-
-import { Remessa } from '../types';
 import {
 	Drawer,
 	DrawerTrigger,
@@ -36,6 +12,8 @@ import {
 } from '@/components/ui/drawer';
 import { useIsMobile } from '@/hooks/use-mobile';
 import { useState } from 'react';
+import { Remessa } from '../../types';
+import { RemessaDetailsSidesheet } from '../remessa-details-sidesheet';
 
 function VerMaisButton({ remessa }: { remessa: Remessa }) {
 	const isMobile = useIsMobile();
@@ -170,7 +148,7 @@ function VerMaisButton({ remessa }: { remessa: Remessa }) {
 	);
 }
 
-const columns: ColumnDef<Remessa>[] = [
+export const remessaColumns: ColumnDef<Remessa>[] = [
 	{
 		accessorKey: 'id',
 		header: 'ID',
@@ -254,114 +232,20 @@ const columns: ColumnDef<Remessa>[] = [
 		},
 	},
 	{
-		id: 'verMais',
-		header: '',
+		id: 'actions',
+		header: 'Ações',
 		cell: ({ row }) => {
 			const remessa = row.original;
-			return <VerMaisButton remessa={remessa} />;
+			return (
+				<div className='flex items-center space-x-2'>
+					<VerMaisButton remessa={remessa} />
+					<RemessaDetailsSidesheet
+						remessa={remessa}
+						mode='edit'
+						onRemessaSaved={() => window.location.reload()}
+					/>
+				</div>
+			);
 		},
 	},
 ];
-
-export function RemessaTable({ data }: { data: Remessa[] }) {
-	const [sorting, setSorting] = React.useState<SortingState>([]);
-	const [columnFilters, setColumnFilters] =
-		React.useState<ColumnFiltersState>([]);
-
-	const table = useReactTable({
-		data,
-		columns,
-		getCoreRowModel: getCoreRowModel(),
-		getPaginationRowModel: getPaginationRowModel(),
-		onSortingChange: setSorting,
-		onColumnFiltersChange: setColumnFilters,
-		getFilteredRowModel: getFilteredRowModel(),
-		getSortedRowModel: getSortedRowModel(),
-		state: {
-			sorting,
-			columnFilters,
-		},
-	});
-
-	return (
-		<div>
-			<div className='flex items-center py-3 rounded-xs justify-between'></div>
-
-			<div className='overflow-hidden rounded-lg border'>
-				<Table>
-					<TableHeader className='bg-muted sticky top-0 z-10'>
-						{table.getHeaderGroups().map((headerGroup) => (
-							<TableRow key={headerGroup.id}>
-								{headerGroup.headers.map((header) => {
-									return (
-										<TableHead
-											key={header.id}
-											colSpan={header.colSpan}
-										>
-											{header.isPlaceholder
-												? null
-												: flexRender(
-														header.column.columnDef
-															.header,
-														header.getContext()
-												  )}
-										</TableHead>
-									);
-								})}
-							</TableRow>
-						))}
-					</TableHeader>
-					<TableBody>
-						{table.getRowModel().rows?.length ? (
-							table.getRowModel().rows.map((row) => (
-								<TableRow
-									key={row.id}
-									data-state={
-										row.getIsSelected() && 'selected'
-									}
-								>
-									{row.getVisibleCells().map((cell) => (
-										<TableCell key={cell.id}>
-											{flexRender(
-												cell.column.columnDef.cell,
-												cell.getContext()
-											)}
-										</TableCell>
-									))}
-								</TableRow>
-							))
-						) : (
-							<TableRow>
-								<TableCell
-									colSpan={columns.length}
-									className='h-24 text-center'
-								>
-									Nenhuma remessa encontrada.
-								</TableCell>
-							</TableRow>
-						)}
-					</TableBody>
-				</Table>
-			</div>
-
-			<div className='flex items-center justify-end space-x-2 py-4'>
-				<Button
-					variant='outline'
-					size='sm'
-					onClick={() => table.previousPage()}
-					disabled={!table.getCanPreviousPage()}
-				>
-					Previous
-				</Button>
-				<Button
-					variant='outline'
-					size='sm'
-					onClick={() => table.nextPage()}
-					disabled={!table.getCanNextPage()}
-				>
-					Next
-				</Button>
-			</div>
-		</div>
-	);
-}
