+++ src/features/remessa/index.ts
@@ -1,5 +1,12 @@
-export { RemessaTable } from './components/remessa-table';
+export { RemessaTable } from './components/tables';
 export { RemessaStatsComponent } from './components/remessa-stats';
 export { NovaRemessaSidebar } from './components/nova-remessa-sidebar';
+export { RemessaDetailsSidesheet } from './components/remessa-details-sidesheet';
+export {
+	RemessaStatusCard,
+	WineTypeCard,
+	QualityRateCard,
+} from './components/cards';
 export { RemessaService } from './services/remessa-service';
+export { useRemessas } from './hooks/useRemessas';
 export type { Remessa, RemessaStats } from './types';
